python/requirements.txt
Flask==3.0.3
flask-cors==4.0.1
python-dotenv==1.0.1

numpy==1.26.4
opencv-python==4.10.0.84
Pillow==10.4.0

torch==2.3.1
torchvision==0.18.1
timm==0.9.16

joblib==1.4.2
scikit-learn==1.5.1

boto3==1.35.10
requests==2.32.3
gunicorn==22.0.0
